requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
bandit==1.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
discord.py==1.6.0
gitdb==4.0.5
GitPython==3.1.13
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
multidict==5.1.0
oauthlib==3.1.0
packaging==20.9
pbr==5.5.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
smmap==3.0.5
stevedore==3.3.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
yarl==1.6.3
